sanic/requirements.txt
aiofiles==0.6.0
alembic==1.4.3
bcrypt==3.2.0
certifi==2020.11.8
cffi==1.14.4
h11==0.9.0
httpcore==0.11.1
httptools==0.1.1
httpx==0.15.4
idna==2.10
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.10.0
multidict==5.0.0
postgres==3.0.0
psycopg2-binary==2.8.6
psycopg2-pool==1.1
pycparser==2.20
PyJWT==2.0.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
rfc3986==1.4.0
sanic==20.9.1
six==1.15.0
sniffio==1.2.0
SQLAlchemy==1.3.22
ujson==4.0.1
websockets==8.1
